Bug 40191: Correct GeckoView version
When fetching the dependencies for Fenix for some reason an older
GeckoView version gets downloaded: the one used immediately before the
final bump for android-components 70.0.0.

I am not sure which .pom file is the culprit but this breaks our builds
when using our android-components 70.0.0 which requires the GeckoView
dependency it got actually tagged with.