Bug 40203: autoflush STDOUT in sign-nightly
We save logs of the sign-nightly script by redirecting STDOUT and STDERR
to a file. However, when the output is not a terminal, STDOUT will have
buffering by default, but STDERR will not, which results in some lines
being written in the wrong order in our log file. To avoid that we
enable autoflush on STDOUT.