Bug 40245: Rebuild all container images
In #40240 we found that package updates were not installed in containers
created using debootstrap, but are now installed in the containers
created with mmdebstrap. Some of those updates are causing differences
in the build of some components.

To be able to build both the stable and master branches from the same
tor-browser-build directory without sharing containers, we update the
filenames of all containers.