Bug 40605: Reworked the macOS toolchain creation.
We were compiling another Clang with the macOS toolchain, but we did not
really need to, so I have removed that step.

However, compiling compiler-rt was failing in this way, so I have
disabled a pair of components we should not need. Probably, only the
built-ins are actually needed.

Finally, we now compile libc++ in the way LLVM currently suggests.