Bug 40626: Define the replacements for generic families on Linux
FontConfig does not know much about our fonts, and falls back to the
font version, when it cannot find any other criterion to decide which
font to use.
So, we want to make sure that at least sans-serif, serif and monospace
are the fonts we want (currently, Arimo, Tinos and Cousine).