Bug 40665: Shorten snowflake bridge line.
The bridge parameters are too long (535 bytes) to fit in space available
for encoding them (510 bytes). This change removes one of the STUN
servers, stun:stun.voip.blackberry.com:3478, to reduce the length to 500
bytes.

stun:stun.voip.blackberry.com:3478 was chosen because, besides being a
relatively long string in itself, it's known to be blocked (on the
Unified Register) in Russia:
https://lists.torproject.org/pipermail/anti-censorship-team/2022-May/000237.html