Bug 40686: Add a temporary project to fetch Fluent tranlations for base-browser
We are adding a notification to tell users they can change their
language. It needs new strings, and I have decided to use Fluent for
them, for a series of reasons.

However, it is the first component to use Fluent, and it needs to be
translated with Weblate, but we are not moving base-browser to Weblate,
yet.
So, we are using another branch on translation.git and we need a new
project to fetch it, which I am adding with this commit.

Also, I have changed how we manage base-browser translations, and
transformed the TT for to a bash for, to keep the Firefox build script
shorter, which is helpful for build debugging sessions.