Bug 40699: Fix input_files in projects/firefox-l10n/config
In input_files in projects/firefox-l10n/config we are doing this:

  my $locales = project_config($project, 'var/locales', $options);
  foreach my $locale (@$locales) {
    $locale = process_template($project, $locale, '.');

In this foreach loop, `$locale` is pointing to the real string from
`var/locales` and not a copy, and we are updating it with the output
from `process_template`. This means that the template will be processed
the first time the function is called (for example for a linux-x86_64
build) and the processed value from the first call reused on subsequent
calls (for example a macos build, where the `ja` locale is supposed to
be different).

This is what caused the reproducibility issue in
https://gitlab.torproject.org/tpo/applications/tor-browser-build/-/merge_requests/591#note_2857042

To fix that we just use a separate variable to store the processed
template.