Bug 40732: Review the Bundle-Data and do not ship the default profile
In this commit we stop shipping Bundle-Data/$osname in base-browser,
which means that the default profile is now distributed only in
Tor Browser (and only on Windows and Linux).
For this reason, we had to modify also the start-browser script.

On macOS we still had profile.ini, which we do not need anymore, so this
commit removes it.
Also, we are disabling the profile migration in the browser build
configuration (tor-browser#41577), so the logic to populate the
override.ini has been removed, too.

Finally, we had two .dmg templates: one for installing Tor Browser to
/Applications, and one to install it to the desktop.
We were not using the latter anymore, and this commits duplicate the
former to start differentiating the dmg resources between the various
browsers.