Bug 28124: Switch to Mozilla's libdmg-hfsplus
To show the DMG icon it seems we need to create the DMG from a HFS
filesystem, rather than an ISO one.
So, to then do so, with this commit I am switching to Mozilla's fork of
libdmg-hfsplus, I am updating its build script and using it to build
also the hfsplus tool.

Also, add the hfsplus project, which is needed to create the HFS
filesystem in the first place.