Bug 40750: Solve rlbox reproducibility problems
The wasi-sandbox project was not reproducible, because libc.a was
filled in a non-deterministic order.
Its Makefile contains some wildcards, but adding a sort to them was not
enough to build libc.a deterministically.

After fixing this problem, Firefox build issue disappears, too.

Upstream issue: https://github.com/WebAssembly/wasi-libc/issues/398