Bug 40847: Do not rely on du to compute the HFS size
HFS has the disadvantage that we cannot grow the filesystem while
adding files.
We have to estimate an initial size, instead, and we relied on du for
this. However, du depends on the underlying filesystem, and this could
lead to reproducibility issues (because the HFS headers tell different
sizes).

So, with this change, we compute a rough number of needed blocks by
taking the actual size of the files.

Overshooting is not a problem, because DMG is compressed with bzip2, so
unused space will be trimmed, eventually.