Bug 40855: Update toolchains for Firefox 115 (common part)
Toolchain updates needed for all platforms.

Bug 31588 (partially): Create cargo vendor archives locally for
cbindgen.

The Rust crates we use provide a Cargo.lock. That should be enough to
make sure they always use a certain version of a dependency.
Therefore, with 115, we could try stop hosting vendor archives on
people.tpo and build them locally.
In this commit we do that for cbindgen.
In next commit we could try to do so for also other projects.