Bug 40855: Update toolchains for Firefox 115 (macOS part)
macOS-only toolchain updates.

With this update, we stop downloading the macOS SDK from people.tpo.
Instead, we download it from apple.com and we use the same scripts as
Mozilla to extract it.
We took them from Firefox source tree, and slightly adapted to our
build system/containers/needs.

Also, we try to be more similar to Mozilla in compiling LLVM runtimes.