Bug 40089: Change how we set MOZ_BUILD_DATE
For normal alpha/release builds, we use the date from the commit, on
which we add the numbers from the version string as seconds (in order to
keep increasing the build date when we release a new version without
adding new commits).

For nightly-testbuild, we use the date from the commit.

For nightly we try to parse the day from the version number, or use the
current day at 00:00:01.