Bug 31546: Copy Firefox PDBs for Windows
Also copy debug symbols after stripping on Linux 32-bit (we only copied
them in Linux 64-bit) and for all our browsers (previously we copied
them only for Tor Browser).

Include the headers directory with the symbols, because some of them
are generated during the build, but they are needed for debugging.