Move the macOS lld patch to llvm-project.
To solve a macOS reproducibility project, we backported a patch from
LLVM, but we applied it only to clang and only to macOS to prevent a
rebuild for all the other platforms.
However, a new rebuild is now needed, so we can apply the patch to all
the platforms and all the project using the LLVM source code for
consistency, even though the previous solution was completely fine.