Bug 40922: Use base-browser.ftl instead of languageNotification.ftl.
We have finally renamed the file for Base Browser strings, and moved it
to the same branch as other Base Browser string files.
So, we needed to change the references to the Fluent file (we have one
in Firefox's build file), and then we could drop the references to the
basebrowser-newidentity branch, since it is not used anymore.