Bug 40924: Customize the remoting name.
Firefox uses the remote name also as GTK's prgname, which is then shown
in some UI elements like GNOME's Alt-Tab.
We used to customize it with --class and --name on Linux, but these
options are lost when restarting the browser after an update.

On other platforms it should not have evident consequences.