Bug 40931: Fix incrementals after #40829
We update tools/signing/nightly to fix signing of nightly incrementals.

We also update tools/update-responses/update_responses to fix the regexp
to match .mar and incremental.mar filenames, since the OS now includes
a `-` (except for macos).

We change the incremental filenames to use `--` instead of `-` to
separate OS and old_version, to make it easier to parse since both of
those can include a `-`.