Bug 40944: In update_responses, make an error if we see an unkwown OS
If a mar file is found with an OS name that doesn't have a corresponding
`build_targets` entry in the config, the OS name was being used as build
target to generate the `.htaccess` file. This is not supposed to happen,
except when something is wrong, so instead we should make an error in
that case.