Bug 40956: Allow testing the updater in the relase and alpha channel.
When override_updater_url is set we used to copy marsigner.der as the
nightly secondary key. With this commit, we check if we are building a
release or an alpha, and in case we replace the secondary release key.

Also, this commit adds a Makefile target to create incrementals and
update response configuration that uses the unsigned files.