Bug 40971: Do not ship Linux i686 debug symbols.
Debug symbols are not actually genereated because we used to have a
linking problem when they were enabled. So, we have been shipping a
tiny package only with generated headers in the last alphas.

Hopefully, this is a temporary change, so I disabled only the final
copy in projects/browser/build, but left the step in
projects/firefox/build enabled.