Bug 40972: Tweaks to the changelog script for MB and group projects
The changelog script works well also with Mullvad Browser, but there
are a few changes we need to do every time we use it for MB instead of
TBB.
With this commit we automate them.

In addition to that, we group items by project, as it makes reading the
changelog easier.