Bug 41016: Switch from bullseye to bookworm on macOS+Windows.
Debian bookworm became the new stable in June 2023, so we should update
our containers to use it.

On macOS the update did not cause any issue, and just updating the
suite name worked.
On Windows, it caused some problems where we used the strip provided by
the OS (only for tor, it seems), because the new version of strip seems
to update the timestamps by default.

We are delaying the process for Android because there are still a
couple of projects that require Java 11, which is not available on
bookworm.