Bug 41015: Enable std::filesystem on libc++ on Windows
We need to do some path manipulation in some Firefox code that is run
before initializing XPCOM.
So, the alternatives are either Path* functions from shlwapi, or
std::filesystem, which is disabled in Firefox 115.

Mozilla enabled it starting from 116, but we have been told it is okay
to enable it also in 115, so we do it with this patch.