Bug 41046: Use the final path in Linux debug symbols.
We had shipped an archive with debug symbols for a long time now.
However, users needed instructions on how to use it, because the files
needed to be moved to be used.
With this commit, it will be possible to just extract the tarball where
the tor-browser directory is located.