Bug 40606: Use Clang to compile NSIS
NSIS officially supports only MSVC and GCC as compilers, so it was one
of the two projects forcing us to keep a GCC toolchain for Windows.
With this commit, we adapt the tricks Mozilla uses to build NSIS with
Clang, and remove the dependency on mingw-w64.

The main trick is to use the GNU assembler instead of Clang's internal
one to build NSIS's system plugin. Luckily, it can be found in
binutils, without a full GCC build.