Bug 41065: Cleanup the NSIS script
Improve the formatting of the installer, and update the language list
to match the one in rbm.conf (even though it would be good to keep the
lists synchronized automatically).

Also, include the channel name in the displayed name and in the default
install directory (except for release).

Since we are changing many lines, I took the ocasion to switch to LF
line endings (NSIS accepts them).