Bug 41077: Draft the new installer for Mullvad Browser.
The installer for Mullvad Browser is going to be different from Tor
Browser's, because it will offer both a traditional installation
(at the moment only for the current user, to avoid enabling privilege
elevation code paths, but still with a control panel entry) and
portable "installations".

The traditional one will support also setting MB as the default
browser and file associations. However, this is quite tricky, and
having a base installer to start implementing the feature is helpful.
Thus this commit, even though the final installer will be likely
changed.