Bug 40569: Include the channel name on macOS dmgs.
This commits make alpha and nightlies macOS build include the channel
in the app bundle.

Also, it renames the Project_Name_Channel to display_name to make it
less ambiguous, and now Project_Name_Channel includes the channel also
for the release target, even though it will be rarely used.