Bug 41089: Register file types in the Mullvad Browser installer.
This step is needed to set Mullvad Browser as a default browser.

Also, we use the same way as Firefox to identify different installs
uniquely based on the hash of their installation directory.
Therefore, we need to build the CityHash NSIS plugin on Firefox and
consume it in browser.
We do the same for the ApplicationID plugin.