Bug 41127: Update default browser_release_date for nightly
In nightly builds, when a date cannot be extracted from
`var/torbrowser_version` (for example when doing a testbuild), we set a
default date. However `firefox-android` fails to build when the date
from `MOZ_BUILD_DATE` (which is based on `var/browser_release_date`) is
too old. So we change the default date to something more recent.