Bug 42535 (TB): Fix Japanese translation paths.
When we setup packaged locales, it seemed that Japanese files needed to
be in torbutton/locale/ja also for macOS, instead of ja-JP-mac.
However, recently we saw that Firefox cannot find brand.ftl in ja, and
the legacy files work also in ja-JP-mac.
Maybe the initial tests were influenced by some cache.