Bug 41149: Add be, bg, and pt-PT to nightlies.
Also, add support for skipping languages when they are evaluated as
empty strings after being processed as templates.
We already process them for Japanese, but with this change we will be
able to enable/disable languages for specific channels.

Also, fix a typo in the "id" language code in the map we use to
associate language names to codes for NSIS.