Bug 41150: Do not check for SSE2 in the Windows installer anymore.
Tor Browser requires SSE2, so a long time ago we added a check on the
Windows installer to warn users in case their machine did not support
this instruction set.
However, Windows 7 is the minimum OS we support, and from 2018 it also
requires SSE2. Moreover, 14.0 will require Windows 10.

We are going to add translations for the installer, and at this point
it makes more sense to remove this check, rather than asking
translators to translate a messages nobody will see in a few months
from now.