Bug 41123: Move firefox-android inside geckoview.
The steps to build Fenix after the migration to the monorepo are very
similar to the ones we used to build it from the previous repository.
However, they need to be run on the tor-browser.git repository, so in
this commit we move part of the firefox-android project to the
merge_aars phase of GeckoView.