Bug 41178: Disable IPv6 in fix_gradle_deps.py.
Our build servers cannot reach maven.mozilla.org through IPv6.

With this commit, we switch to requests (we have it anyway, since we
use it to build changelogs), which downloads dependencies on memory
first, but usually dependencies are small and you need a giant amount
of memory to build Java stuff anyway.