Bug 41232: Implement single-arch testbuilds on the monorepo.
The trick of the fake fat AAR does not work anymore, as the fat AAR is
consumed in the same script where it is created.
However, running an Android build from a single-arch tree seems not to
involve different commands, so we can create a shared script and
include it in both the build steps.