Bug 41236: Remove binutils when not needed.
Some of our projects have binutils as a dependency, even though it
seems they do not need it, or they need it only when building for
Linux.
And this prevents us from having a single Rust toolchain that handles
Windows i686 and x86_64, for example, even though Rust itself can
target both.