Bug 41247: Simplify tools/update-responses/update_responses
Most commands implemented in update_responses expect the channel to be
an optional argument, doing the operation on all channels if no channel
is provided. However we don't really need that feature. In this commit
we simplify the script by making the channel a required argument.

At the same time we remove the get_channel_version command, which was
not used anywhere.