Bug 41247: Drop sha512 hash from update_responses files (except for nightly)
With tor-browser#42737, the sha512 hash included in the update_responses
files is not checked anymore. Since we're doing a watershed update,
support for older versions checking the hash is not needed.

However since we don't do watershed updates in nightly, we'll keep the
sha512 hash in nightly for a little more time.