Bug 40996: Move .nobackup to the root of the repository.
The .nobackup files prevent from creating symlinks between clones,
which is useful for several use cases (multiple clones, directories
in different drives, etc...).

A .nobackup in the root will not interfere with symlinks anymore, and
we think it is acceptable, as a repo will probably be in sync with
GitLab or something similar.
git update-index --assume-unchanged can help if someone wants to
remove the .nobackup.