Bug 41281: Remove `fetch: 1` from nightly target
With rbm#40079, when we have `fetch: if_needed` (the default), rbm will
now do a fetch when `git_hash` is pointing to a branch. The reason for
setting `fetch: 1` in the nightly target was that we are using some
git branches, which were not fetched automatically with `fetch:
if_needed`.