Bug 41386: Upgrade Go to 1.23.x (except on macOS).
We have been using Go 1.22.x for a while, since 1.23 raised the minimum
required version of macOS to 11, and we still support macOS 10.15.
However, 1.22.x is now EOL, so we need to update to a supported series.

1.24.0 would require an additional version on our go bootstrapping
toolchain, so I decided to stick to 1.23.x for now.

This commit also updates the relprep.py script to update only the hash
of the main version (which is referred to indirectly with variables).