Bug 41393 - Create special clang packaged for Linux
Building Firefox requires compiler-rt wasm32 library to be
packaged alongside clang.

We were doing this step inside the firefox and geckoview
projects, but that meant one can't simply download the clang
toolchain and use it for dev builds.

MacOS and Windows already have separate projects to package
clang with other libraries, so a step to include the wasm32
compiler-rt lib was added to each of them.