Bug 41444: Fix macos and windows artifact generation
There were multiple issues.

For MacOS there were missing tools, namely hfsplus and libdmg,
the location of some bundle files were different than expected
and teh generated artifacts were not the same as the ones in
Linux, so we were copying the wrong things. Finally, we were not
taking into account universal builds when copying artifacts.

For Windows we were just not copying the right artifacts.