Bug 26408: Remove redundant check of macos code signature in update_responses
When CHECK_CODESIGNATURE_EXISTS is set (we set it when re-generating
incrementals for macos during the signing process), we were checking
that both the mar files from previous and new versions are code signed.
Checking that the mar file from the previous version is code signed is
not necessary since it is checked just before even when the
CHECK_CODESIGNATURE_EXISTS is not set.