Bug 41604: Keep update-responses files from previous release
We were keeping update-responses directories from the previous release
if they are not in the new one, which happens when a release is not for
all platforms.

We were also keeping the `download-*.json` files. However, in the case
of android-only releases we also need to keep `downloads.json` and
`.htaccess`.

To simplify we now keep all files and directories from the previous
release which are not in the new one, including files starting with a
dot (such as `.htaccess`).