Bug 41580: Updated go to 1.25.3.
This is a major Go update, which introduces incompatibilities.
For this reason, we needed to update also go-licenses.
However, the new version does not build with the Go compiler built to
output Windows binaries (and possibly Android as well).
Therefore, we changed the projects that use it to use the Linux
arfiacts of the tool.