Bug 41627: Build OpenSSL in the Python package.
Build OpenSSL with Python, so that the ssl module can be imported.
We do not use the openssl project to avoid re-building a consistent
part of the Linux toolchain every time we bump it.

Also, bump the Python version since we are at it (we downgraded to be
able to build it with OpenSSL 1.1.x, now we don't have that requirement
anymore), and bump ninja, as the old version didn't build with our new
version of Python.