Bug 41669: Restore UglifyJS but for Android x86_64.
In 15.0 we use UglifyJS for Android x86.
We dropped it in the 16.0 series, as we do not support Android x86, but
with this commit we temporarily restore it for x86_64 to fit in the
Play Store requirements.

With Firefox 148 we might be able to use the same minifier Mozilla
uses, since they recently switched to a different one.